Separate JS projects from JVM projects; make shared a cross-project.
`shared` was used as if it were a cross-project, being depended on
by JVM projects as well as JS projects. This was very bad, as it
introduced a dependency on scalajs-library.jar from the JVM
projects.

We now clearly separate the JS projects from the JVM projects, and
`shared` is made a cross-project, whose JS and JVM parts are used
appropriately.